requirements.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.1
bentoml==1.3.16
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
cattrs==23.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.1.0
cohere==5.13.3
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepmerge==2.0
Deprecated==1.2.15
distro==1.9.0
docx2txt==0.8
dulwich==0.22.6
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.6
fastavro==1.9.7
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
httpx-ws==0.7.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.3.10
langchain-chroma==0.1.4
langchain-cohere==0.3.3
langchain-community==0.3.10
langchain-core==0.3.24
langchain-experimental==0.3.3
langchain-ollama==0.2.1
langchain-text-splitters==0.3.2
langgraph==0.2.59
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.44
langserve==0.3.0
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.18.4
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
oauthlib==3.2.2
ollama==0.4.4
onnxruntime==1.20.1
openai==1.55.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pathlib==1.0.1
pathspec==0.12.1
pillow==11.0.0
pip-requirements-parser==32.0.1
plum-dispatch==1.7.4
posthog==3.7.4
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pyaml==24.12.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
shellingham==1.5.4
simple-di==0.1.5
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
Spire.Doc==12.12.0
SQLAlchemy==2.0.36
sse-starlette==1.8.2
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.5.8
uvicorn==0.32.1
watchdog==6.0.0
watchfiles==1.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
